recognition_ai/requirements.txt
openai-whisper
fastapi
uvicorn
python-multipart
pdfplumber
pytesseract
Pillow